pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copytalker"
dynamic = ["version"]
description = "Cross-modal data conversion driven asynchronous multi-voice translation system"
readme = "README.md"
requires-python = ">=3.9"
license = "GPL-3.0-only"
authors = [
    {name = "CopyTalker Team", email = "copytalker@example.com"}
]
maintainers = [
    {name = "CopyTalker Team", email = "copytalker@example.com"}
]
keywords = [
    "speech-to-speech",
    "translation",
    "text-to-speech",
    "speech-recognition",
    "multi-language",
    "real-time",
    "audio-processing",
    "whisper",
    "kokoro-tts",
    "nllb",
    "voice-cloning",
    "indextts",
    "fish-speech",
    "emotion-tts",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Natural Language :: Chinese (Simplified)",
    "Natural Language :: Japanese",
    "Natural Language :: Korean",
    "Natural Language :: French",
    "Natural Language :: German",
    "Natural Language :: Spanish",
    "Natural Language :: Russian",
    "Natural Language :: Arabic",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "faster-whisper>=0.6.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "sounddevice>=0.4.4",
    "webrtcvad-wheels>=2.0.10",
    "numpy>=1.26.0",
    "huggingface-hub>=0.16.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "sentencepiece>=0.1.96",
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
    # TTS engines - fast to install
    "edge-tts>=6.1.0",
    "pyttsx3>=2.90",
    "fish-audio-sdk>=0.1.0",
    "httpx>=0.24.0",
    # GUI
    "PySide6>=6.5.0",
    # Python 3.13 removed audioop; pydub needs this shim
    "audioop-lts>=0.2.1; python_version>='3.13'",
]

[project.optional-dependencies]
# Kokoro high-quality neural TTS (requires model download)
kokoro = [
    "kokoro>=0.1.0",
    "librosa>=0.9.1",
]

# CJK language support (Chinese, Japanese, Korean) - recommended for Asian languages
cjk = [
    "cn2an>=0.5.17",
    "pypinyin>=0.44.0",
    "pypinyin-dict>=0.1.0",
    "jieba>=0.42.1",
    "fugashi>=1.1.0",
    "jaconv>=0.3.4",
    "mojimoji>=0.0.12",
    "unidic-lite>=1.0.0",
]

# IndexTTS voice cloning engine (not available on macOS)
indextts = [
    "pynini>=2.1.5",
]

# PyAudio backend (alternative to sounddevice)
pyaudio = [
    "pyaudio>=0.2.11",
]

# Development dependencies
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-timeout>=2.1.0",
    "pytest-xdist>=3.3.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "ruff>=0.0.270",
    "types-PyYAML>=6.0.0",
]

# Documentation
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-autodoc-typehints>=1.23.0",
]

# Complete installation with Kokoro and CJK support
complete = [
    "copytalker[kokoro,cjk]",
]

# All optional dependencies including development
all = [
    "copytalker[kokoro,cjk,indextts,pyaudio,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/cycleuser/CopyTalker"
Repository = "https://github.com/cycleuser/CopyTalker.git"
Documentation = "https://github.com/cycleuser/CopyTalker#readme"
Issues = "https://github.com/cycleuser/CopyTalker/issues"
Changelog = "https://github.com/cycleuser/CopyTalker/blob/main/CHANGELOG.md"

[project.scripts]
copytalker = "copytalker.cli.main:main"
copytalker-gui = "copytalker.gui.app:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["copytalker", "copytalker.core", "copytalker.audio", "copytalker.speech", "copytalker.translation", "copytalker.tts", "copytalker.gui", "copytalker.gui.views", "copytalker.gui.widgets", "copytalker.gui.controllers", "copytalker.gui.qt", "copytalker.gui.qt.views", "copytalker.gui.qt.widgets", "copytalker.gui.qt.controllers", "copytalker.cli", "copytalker.utils"]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "copytalker.__version__"}

[tool.setuptools.package-data]
copytalker = [
    "py.typed",
    "data/*.json",
]

# ============ Tool Configurations ============

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true
strict_optional = true
show_error_codes = true

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function call in default argument
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests as requiring GPU",
    "audio: marks tests as requiring audio hardware",
    "integration: marks tests as integration tests",
]
timeout = 300
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["copytalker"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__main__.py",
    "*/gui/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "@abstractmethod",
]
show_missing = true
